+++ WPF/WpfDockApp/ChargeOfPartyUserControl.xaml
@@ -5,7 +5,8 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:WpfDockApp"
              mc:Ignorable="d" 
-             Width="330" Height="140" FontSize="16">
+             Width="330" Height="140" FontSize="16"
+        d:DataContext="{d:DesignInstance local:ChargeOfPartyUserControl}">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition/>
@@ -16,8 +17,8 @@
             <ColumnDefinition/>
         </Grid.ColumnDefinitions>
         <TextBlock Grid.Column="0" Grid.Row="0" Text="Name" VerticalAlignment="Center" Margin="20,0,0,0"/>
-        <TextBox Grid.Column="1" Grid.Row="0" Width="200" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="20,0,0,0"/>
+        <TextBox Grid.Column="1" Grid.Row="0" Width="200" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="20,0,0,0" Text="{Binding UserName}"/>
         <TextBlock Grid.Column="0" Grid.Row="1" Text="Tel" VerticalAlignment="Center" Margin="20,0,0,0"/>
-        <TextBox Grid.Column="1" Grid.Row="1" Width="200" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="20,0,0,0"/>
+        <TextBox Grid.Column="1" Grid.Row="1" Width="200" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="20,0,0,0" Text="{Binding Tel}"/>
     </Grid>
 </UserControl>
