+++ WPF/WpfDockApp/ReservationInformationUserControl.xaml
@@ -18,10 +18,10 @@
             <ColumnDefinition Width="160"/>
             <ColumnDefinition/>
         </Grid.ColumnDefinitions>
-        <CheckBox Grid.Column="0" Grid.Row="0" Content="Smoking" Margin="20,0,0,0" FontSize="16"/>
+        <CheckBox Grid.Column="0" Grid.Row="0" x:Name="Smoking" Content="Smoking" Margin="20,0,0,0" FontSize="16"/>
         <TextBlock Grid.Column="0" Grid.Row="1" Text="Number of people" VerticalAlignment="Center" Margin="20,0,0,0"/>
-        <TextBox Grid.Column="1" Grid.Row="1" Width="200" VerticalAlignment="Center" Margin="20,0,0,0" HorizontalAlignment="Left"/>
-        <RadioButton Grid.Column="0" Grid.Row="2" Content="Course" Margin="20,0,0,0" Background="{x:Null}" FontSize="16"/>
-        <RadioButton Grid.Column="0" Grid.Row="3" Content="A la carte" Margin="20,0,0,0" Background="{x:Null}" FontSize="16"/>
+        <TextBox Grid.Column="1" Grid.Row="1" x:Name="NumberOfPeople" Width="200" VerticalAlignment="Center" Margin="20,0,0,0" HorizontalAlignment="Left"/>
+        <RadioButton Grid.Column="0" Grid.Row="2" x:Name="Course" Content="Course" Margin="20,0,0,0" Background="{x:Null}" FontSize="16"/>
+        <RadioButton Grid.Column="0" Grid.Row="3" x:Name="Alacarte" Content="A la carte" Margin="20,0,0,0" Background="{x:Null}" FontSize="16"/>
     </Grid>
 </UserControl>
