+++ WPF/WpfDockApp/TreeUserControl.xaml
@@ -1,0 +1,13 @@
+﻿<UserControl x:Class="WpfDockApp.TreeUserControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:WpfDockApp"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <TreeView x:Name="TreeView">
+        </TreeView>
+    </Grid>
+</UserControl>
