+++ WPF/WpfDockApp/WpfDockApp.csproj
@@ -142,6 +142,7 @@
     <Compile Include="CustomControlWindow.xaml.cs">
       <DependentUpon>CustomControlWindow.xaml</DependentUpon>
     </Compile>
+    <Compile Include="MakeDocumentEventArgs.cs" />
     <Compile Include="MultiUserControlWindow.xaml.cs">
       <DependentUpon>MultiUserControlWindow.xaml</DependentUpon>
     </Compile>
@@ -154,6 +155,9 @@
     <Compile Include="SearchEventArgs.cs" />
     <Compile Include="SimpleWindow.xaml.cs">
       <DependentUpon>SimpleWindow.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="TreeUserControl.xaml.cs">
+      <DependentUpon>TreeUserControl.xaml</DependentUpon>
     </Compile>
     <Page Include="ChargeOfPartyUserControl.xaml">
       <SubType>Designer</SubType>
@@ -188,6 +192,10 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="SimpleWindow.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="TreeUserControl.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
