+++ WPF/WpfDockApp/packages.config
@@ -1,8 +1,15 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Codeer.Friendly" version="2.6.1" targetFramework="net472" />
+  <package id="Codeer.Friendly.Windows" version="2.15.0" targetFramework="net472" />
+  <package id="Codeer.Friendly.Windows.Grasp" version="2.12.0" targetFramework="net472" />
+  <package id="Codeer.Friendly.Windows.NativeStandardControls" version="2.15.0" targetFramework="net472" />
+  <package id="Codeer.TestAssistant.GeneratorToolKit" version="3.9.0" targetFramework="net472" />
   <package id="ControlzEx" version="4.3.1" targetFramework="net472" />
   <package id="Extended.Wpf.Toolkit" version="4.0.1" targetFramework="net472" />
   <package id="MahApps.Metro" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.Xaml.Behaviors.Wpf" version="1.1.19" targetFramework="net472" />
+  <package id="Ong.Friendly.FormsStandardControls" version="2.16.0" targetFramework="net472" />
+  <package id="RM.Friendly.WPFStandardControls" version="1.40.0" targetFramework="net472" />
   <package id="Xceed.Wpf.AvalonDock.Themes.VS2013" version="2.9.0.1" targetFramework="net472" />
 </packages>
